Atividades da Primeira aula de C#

+++ aula1exercicios/Atividade 2/Program.cs
@@ -4,7 +4,7 @@ namespace Aula1 {
     class Program {
         static void Main (string[] args) {
 
-            int num1;
+            float num1;
             string Fulano;
             float resultado;
 
@@ -13,10 +13,8 @@ namespace Aula1 {
 
             Fulano = Console.ReadLine ();
 
-            // Concatenação
             Console.WriteLine (Fulano + ", Agora digite o valor que quer consultar: ");
 
-            // Converto o texto capturado para o tipo inteiro
             num1 = int.Parse (Console.ReadLine ());
 
             resultado = num1 * 30 / 100 + num1;
